# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktaba Dot Com - Arabic RTL E-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponApplyInput"
      responses:
        "200":
          description: Cart with coupon applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCoupon
      tags: [cart]
      responses:
        "200":
          description: Cart without coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── ORDERS (PUBLIC) ─────────────────────────────────────────────────────────
...
  /admin/coupons:
    get:
      operationId: adminListCoupons
      tags: [admin]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponListResponse"
    post:
      operationId: adminCreateCoupon
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /admin/coupons/{id}:
    patch:
      operationId: adminUpdateCoupon
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: adminDeleteCoupon
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── ADMIN – SHIPPING ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CartItem:
      type: object
      required: [productId, quantity, unitPrice, subtotal]
      properties:
        productId: { type: integer }
        nameAr: { type: string }
        coverImage: { type: ["string", "null"] }
        slug: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        oldPrice: { type: ["number", "null"] }
        subtotal: { type: number }
        inStock: { type: boolean }
        stockQuantity: { type: integer }
        freeShipping: { type: boolean }
        freeShippingBadgeText: { type: ["string", "null"] }
    Cart:
      type: object
      required: [items, subtotal, shippingCost, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        discount: { type: number }
        shippingCost: { type: number }
        total: { type: number }
        couponCode: { type: ["string", "null"] }
        couponDiscount: { type: number }
        governorateId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        freeShippingReason: { type: ["string", "null"] }
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }
    CouponApplyInput:
      type: object
      required: [code]
      properties:
        code: { type: string }

    # ── Orders ───────────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, type, value, isActive]
      properties:
        id: { type: integer }
        code: { type: string }
        type: { type: string, enum: [percentage, fixed, free_shipping] }
        value: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        maxUses: { type: ["integer", "null"] }
        usedCount: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
    CouponInput:
      type: object
      required: [code, type, value]
      properties:
        code: { type: string }
        type: { type: string, enum: [percentage, fixed, free_shipping] }
        value: { type: number }
        minOrderAmount: { type: number }
        maxUses: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
    CouponUpdate:
      type: object
      properties:
        code: { type: string }
        type: { type: string }
        value: { type: number }
        minOrderAmount: { type: number }
        maxUses: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        isActive: { type: boolean }
    CouponListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Coupon"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Admin Products ────────────────────────────────────────────────────────
...